Split classes into new files
