+++ dataset.py
@@ -3,74 +3,18 @@ import nltk
 import pandas as pd
 import torch
 from torch.utils.data import Dataset, DataLoader
-from torch.nn.utils.rnn import pad_sequence
+
 from PIL import Image
 from torchvision import transforms
 
-
-class Vocabulary():
-    """
-    Class for building and handling vocabulary in dataset
-    """
-    def __init__(self):
-        #initialize known tokens
-        self.index_to_word = {0: "<PAD>", 1:"<START>", 2:"<END>",3:"<UNK>"}
-
-        self.word_to_index = {v:k for k, v in self.index_to_word.items()}
-
-    def __len__(self):
-        return len(self.word_to_index)
-
-    @staticmethod
-    def tokenize(text):
-        return nltk.word_tokenize(text.lower())
-    
-    def build_vocabulary(self,sentence_list,reduce=True, max_size=5000):
-        #download punctuation
-        nltk.download("punkt")
-        freq = {}
-
-        #Build vocabulary set on word frequency
-        for sentence in sentence_list:
-            for word in self.tokenize(sentence):
-                if word in freq:
-                   freq[word] += 1
-                else:
-                    freq[word] = 1         
-                
-        #Sort by frequency
-        items=list(freq.items())
-        items.sort(key=lambda x:x[1],reverse=True)
-
-        #if reduce is True, keep only max_size words in freq order
-        if reduce: 
-            items=items[:max_size]
-            
-        idx = len(self.index_to_word) #starting idx
-
-        #Fill up itow to add it to initial dictionary
-        for k,v in items:
-            self.index_to_word[idx]=k
-            idx+=1
-        self.word_to_index = {v:k for k, v in self.index_to_word.items()}
-        
-    #to assign token from vocab from text
-    def numericalize(self,text):
-        tok_text = self.tokenize(text)
-        return [self.word_to_index[token] if token in self.word_to_index else self.word_to_index['<UNK>'] for token in tok_text]
-
-    #to reconstruct caption
-    #TODO: Pending to concat caption
-    def generate_caption(self,vec):
-        return [self.index_to_word[token] if token in self.index_to_word else '<UNK>' for token in vec.tolist()]
-
- 
+from vocabulary import Vocabulary
+from caps_collate import CapsCollate
 class Flickr8kDataset(Dataset):
     """
     Dataset for Flickr8k data treatment
     """
 
-    def __init__(self, dataset_folder='./dataset',transform=None,reduce=False,max_size=5000):
+    def __init__(self, dataset_folder='/local/Datasets/8KFlicker/archive',transform=None,reduce=False,max_size=5000):
         super().__init__()
         self.transform = transform
         self.images_folder = os.path.join(dataset_folder,'Images')
@@ -122,21 +66,7 @@ class Flickr8kDataset(Dataset):
         tok_vec += [self.vocab.word_to_index['<END>']]
 
         return image, torch.tensor(tok_vec)
-class CapsCollate:
-    """
-    Collate to apply the padding to the captions with dataloader
-    """
-    def __init__(self,pad_idx,batch_first=False):
-        self.pad_idx = pad_idx
-        self.batch_first = batch_first
-    
-    def __call__(self,batch):
-        imgs = [item[0].unsqueeze(0) for item in batch]
-        imgs = torch.cat(imgs,dim=0)
-        
-        targets = [item[1] for item in batch]
-        targets = pad_sequence(targets, batch_first=self.batch_first, padding_value=self.pad_idx)
-        return imgs,targets
+
 
 if __name__ == "__main__":
     image_size = (128,128)
